single app page rehaul
